requirements.txt
virtualenv>=20.23.0
pandas>=2.*
openpyxl>=3.1.2
apache-airflow>=2.6.1